feat: 7일 연속 과제를 전부 완료했는지 확인하는 기능 구현

+++ src/main/kotlin/com/barostartbe/domain/assignment/repository/AssignmentRepository.kt
@@ -2,6 +2,7 @@ package com.barostartbe.domain.assignment.repository
 
 import com.barostartbe.domain.assignment.entity.Assignment
 import com.barostartbe.domain.assignment.entity.enum.AssignmentStatus
+import com.barostartbe.domain.assignment.entity.enum.Subject
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
@@ -52,5 +53,5 @@ interface AssignmentRepository : JpaRepository<Assignment, Long> {
         ) streaks
     """, nativeQuery = true
     )
-    fun findMaxConsecutivePerfectDays(@Param("menteeId") menteeId: Long): Long
+    fun findMaxConsecutivePerfectDays(@Param("menteeId") menteeId: Long): Int
 }
